Move logging utility within `Logger` class
