+++ proxy/common/flag.py
@@ -20,13 +20,14 @@ import multiprocessing
 from typing import Optional, List, Any, cast
 
 from .types import IpAddress
-from .utils import text_, bytes_, setup_logger, is_py2, set_open_file_limit
+from .utils import text_, bytes_, is_py2, set_open_file_limit
 from .utils import import_plugin, load_plugins
 from .constants import COMMA, DEFAULT_DATA_DIRECTORY_PATH, DEFAULT_NUM_WORKERS
 from .constants import DEFAULT_DEVTOOLS_WS_PATH, DEFAULT_DISABLE_HEADERS, PY2_DEPRECATION_MESSAGE
 from .constants import PLUGIN_DASHBOARD, PLUGIN_DEVTOOLS_PROTOCOL
 from .constants import PLUGIN_HTTP_PROXY, PLUGIN_INSPECT_TRAFFIC, PLUGIN_PAC_FILE
 from .constants import PLUGIN_WEB_SERVER, PLUGIN_PROXY_AUTH
+from .logger import Logger
 
 from .version import __version__
 
@@ -94,7 +95,11 @@ class FlagParser:
             sys.exit(1)
 
         # Discover flags from requested plugin.
-        # This also surface external plugin flags under --help
+        # This will also surface external plugin flags
+        # under --help.
+        #
+        # TODO: --plugin(s) can either be a comma separated
+        # list of plugins or could be a list of plugins.
         for i, f in enumerate(input_args):
             if f in ('--plugin', '--plugins'):
                 import_plugin(bytes_(input_args[i + 1]))
@@ -108,7 +113,7 @@ class FlagParser:
             sys.exit(0)
 
         # Setup logging module
-        setup_logger(args.log_file, args.log_level, args.log_format)
+        Logger.setup_logger(args.log_file, args.log_level, args.log_format)
 
         # Setup limits
         set_open_file_limit(args.open_file_limit)
